Making it work with 64bit for D64
I changed this because DWORD etc..  cause issues with D64's shootable switches, shootable secrets and other things.  It seems to be incredibly sensitive to the width of the type used, so I used these, which are still standardised Windows'isms also.

+++ src/engine/i_w3swrapper.h
@@ -45,9 +45,9 @@ typedef bool dboolean;
 
 #ifdef _WIN32
 #include <Windows.h>
-typedef BYTE			byte;
-typedef WORD			word;
-typedef DWORD			dword;
+typedef UINT8			byte;
+typedef UINT16			word;
+typedef UINT64			dword;
 #else
 typedef uint8_t         byte;
 typedef uint16_t		word;
